UIC-ids568-milestone1/GCP/requirements.txt
scikit-learn==1.3.2
joblib==1.3.2
google-cloud-storage==2.10.0